refactor: move to typescript, drop babel and flow
BREAKING CHANGE:

Flow is no longer supported in this project. This project only uses
TypeScript now.
