+++ index.ts
@@ -1,25 +1,36 @@
-/* @flow strict */
-
-type ThrottleOptions = {|
-  start?: boolean,
-  middle?: boolean,
+interface ThrottleOptions {
+  /**
+   * Fire immediately on the first call.
+   */
+  start?: boolean
+  /**
+   * Fire as soon as `wait` has passed.
+   */
+  middle?: boolean
+  /**
+   * Cancel after the first successful call.
+   */
   once?: boolean
-|}
-export function throttle<T: $ReadOnlyArray<mixed>>(
-  callback: (...T) => mixed,
-  wait: number = 0,
+}
+
+interface Throttler<T extends unknown[]> {
+  (...args: T): void
+  cancel(): void
+}
+
+export function throttle<T extends unknown[]>(
+  callback: (...args: T) => unknown,
+  wait = 0,
   {start = true, middle = true, once = false}: ThrottleOptions = {}
-): (...T) => void {
+): Throttler<T> {
   let last = 0
-  let timer
+  let timer: number | undefined
   let cancelled = false
-  const fn = (...args) => {
+  const fn = (...args: T) => {
     if (cancelled) return
     const delta = Date.now() - last
     last = Date.now()
     if (start) {
-      //eslint-disable-next-line flowtype/no-flow-fix-me-comments
-      // $FlowFixMe this isn't a const
       start = false
       callback(...args)
       if (once) fn.cancel()
@@ -42,10 +53,10 @@ export function throttle<T: $ReadOnlyArray<mixed>>(
   return fn
 }
 
-export function debounce<T: $ReadOnlyArray<mixed>>(
-  callback: (...T) => mixed,
-  wait: number = 0,
+export function debounce<T extends unknown[]>(
+  callback: (...args: T) => unknown,
+  wait = 0,
   {start = false, middle = false, once = false}: ThrottleOptions = {}
-): (...T) => void {
+): Throttler<T> {
   return throttle(callback, wait, {start, middle, once})
 }
